+++ apheleia-utils.el
@@ -35,6 +35,7 @@ always returns nil to defer to the formatter."
     (unless indent-var
       (setq indent-var
             (cl-case major-mode
+              (cperl-mode 'cperl-indent-level)
               (css-mode 'css-indent-offset)
               (css-ts-mode 'css-indent-offset)
               (js-jsx-mode 'js-indent-level)
@@ -47,6 +48,7 @@ always returns nil to defer to the formatter."
               (json-ts-mode 'json-ts-mode-indent-offset)
               (nxml-mode 'nxml-child-indent)
               (robot-mode 'robot-mode-basic-offset)
+              (perl-mode 'perl-indent-level)
               (scss-mode 'css-indent-offset)
               (web-mode 'web-mode-indent-style)
               (tsx-ts-mode 'typescript-ts-mode-indent-offset)